+++ x/posts/client/rest/rest.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/types/rest"
+	"github.com/desmos-labs/desmos/x/posts/internal/types"
 	"github.com/gorilla/mux"
 )
 
@@ -23,6 +24,7 @@ type CreatePostReq struct {
 	Subspace       string            `json:"subspace"`
 	OptionalData   map[string]string `json:"optional_data"`
 	CreationTime   time.Time         `json:"creation_time"`
+	Medias         types.PostMedias  `json:"medias,omitempty"`
 }
 
 // AddReactionReq defines the properties of a reaction adding request's body.