+++ x/posts/internal/keeper/keeper.go
@@ -55,8 +55,8 @@ func (k Keeper) SavePost(ctx sdk.Context, post types.Post) {
 	store.Set([]byte(types.PostStorePrefix+post.PostID.String()), k.Cdc.MustMarshalBinaryBare(&post))
 
 	// Set the last post id only if the current post has a greater one than the last one stored
-	if post.PostID > k.GetLastPostID(ctx) {
-		store.Set([]byte(types.LastPostIDStoreKey), k.Cdc.MustMarshalBinaryBare(&post.PostID))
+	if id := post.PostID; id > k.GetLastPostID(ctx) {
+		store.Set([]byte(types.LastPostIDStoreKey), k.Cdc.MustMarshalBinaryBare(&id))
 	}
 
 	// Save the comments to the parent post, if it is valid