+++ x/posts/legacy/v0.1.0/types.go
@@ -24,7 +24,7 @@ type Post struct {
 	LastEdited        sdk.Int        `json:"last_edited"`        // Block height at which the post has been edited the last time
 	AllowsComments    bool           `json:"allows_comments"`    // Tells if users can reference this PostID as the parent
 	ExternalReference string         `json:"external_reference"` // Used to know when to display this post
-	Owner             sdk.AccAddress `json:"owner"`              // Creator of the Post
+	Owner             sdk.AccAddress `json:"owner"`              // Creator of the post
 }
 
 // Reaction is a struct of a user like